Moved transform functions to new file
